+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 - Delete embedded .travis-opam.sh (#2)
+- Add `Record.declare0` (#4)
 
 ## v0.2.0